Change the HTML tag and attribute of some elements
In order to access the element of <main> in the javascript, the id="card-list" was introduced. This makes it easier to reference the element.

The <input> element was replaced with the element <textarea> because <input> can only have 1 line. An error message was also introduced to ensure it meets the condition of the javascript program.

The <button> id was renamed from "new-post" to "create-post" to make things much clearer and easier to understand.