Fix bugs, naming issue, etc.
I renamed the routing url because they were inconsistent. I've also fixed the naming issue where it is passing the variable image, instead of profile_image with the value of req.session.pfp instead of the previous one.

I've also removed commentEdit variable and query because I deemed them unnecessary because you could've just use one query instead of two.

I've also done a lot of variable renaming, they're not that big of a deal, so it is not a concern for other participants in this project.

I've added a middleware function named "permissionRequired". It is used to determined if the user is authorized to edit or delete a comment, a post, and it's own profile.